services/infra-runtime/requirements.txt
fastapi==0.104.0
uvicorn==0.23.2
aio-pika==9.3.1
redis==5.0.1
motor==3.6.0
networkx==3.2.1
pydantic-settings==2.0.3
structlog==23.2.0
